telegram_client/pyproject.toml
[tool.poetry]
name = "junction"
version = "0.1.0"
description = ""
authors = ["evgenii <evgeniimigaev@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
python-telegram-bot = "^20.6"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
